Website/satcard/requirements.txt
Django==4.1.6
djangorestframework==3.14.0
mysqlclient
gunicorn
django-filter==22.1
django-allauth==0.49.0
django-crispy-forms==1.14.0
djangorestframework-csv==2.1.1
pandas==2.0.2
numpy==1.24.3
tensorflow==2.13.0rc1
scikit-learn==1.2.2
scipy==1.10.1
emd==0.6.2
django-import-export==3.2.0
django-admin-rangefilter==0.10.0
